python/server/requirements.txt
pyzmq>=25.0.0
msgpack>=1.0.0
